fix/bug-4 and fix feature to prohibit URL duplication
