+++ src/utils/dateUtils.ts
@@ -5,5 +5,5 @@ export const getCurrentJSTTime = (): string => {
   const now = new Date();
   const jstOffset = 9 * 60; // JSTはUTC+9時間
   const jstTime = new Date(now.getTime() + jstOffset * 60 * 1000);
-  return jstTime.toISOString().replace('Z', '+09:00');
+  return jstTime.toISOString()
 };
